pyproject.toml
[tool.poetry]
name = "amazing-hunting"
version = "0.1.0"
description = ""
authors = ["wildtsk <wildtsk@gmail.com>"]
readme = "README.md"
packages = [{include = "amazing_hunting"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
